Change aria-pressed into aria-expanded for buttons connected to expandable content (#1679)
This PR changes `aria-pressed` to `aria-expanded` on buttons toggling expandable content.

Related issue: https://github.com/just-the-docs/just-the-docs/issues/1680.

`aria-pressed` is used for e.g. switch buttons, a screen reader will announce "on" or "off" with the button. [MDN on aria-pressed](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Reference/Attributes/aria-pressed).

For expandable content like menus the correct attribute is `aria-expanded`. The screen reader will announce the accessible name (the aria-label in this case) of the button adding the state expanded or collapsed when the button gets focus. [MDN on aria-expanded](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Reference/Attributes/aria-expanded).

This PR also simplifies aria-label for buttons with `aria-expanded`: when `aria-expanded` is used there is no need to add to the aria-label the text “toggle", because the `aria-expanded` already gives that info plus the state of the toggle.

So: “Menu”, instead of “Toggle menu”.
The screen reader will announce the info as:
> Button Menu collapsed

and

> Button Menu expanded

Also it is for voice recognition software advisable to start the aria-label with the text a user may guess what the button stands for, to focus it.